Application structure modularised
