+++ .gitignore
@@ -1,2 +1,3 @@
+*.sqlite
 __pycache__
 venv/
