+++ app/templates/base.html
@@ -16,11 +16,11 @@
 				<span class="icon-bar"></span>
 				<span class="icon-bar"></span>
 			</button>
-			<a class="navbar-brand" href="/">Blogly</a>
+            <a class="navbar-brand" href="{{ url_for('main.index') }}">Blogly</a>
 		</div>
 		<div class="navbar-collapse collapse">
 			<ul class="nav navbar-nav">
-				<li><a href="/">Home</a></li>
+                <li><a href="{{ url_for('main.index') }}">Home</a></li>
 			</ul>
 		</div>
 	</div>
