--- app/templates/user.html
@@ -1,7 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Blogly{% endblock %}
-{% block page_content %}
-<div class="page-header">
-	<h1>Hello, {{ name }}!</h1>
-</div>
-{% endblock %}
