+++ config.py
@@ -2,7 +2,7 @@ import os
 basedir = os.path.abspath(os.path.dirname(__file__))
 
 class Config:
-    SECRET_KEY = os.environ.get('SECRECT_KEY') or 'hard to guess' 
+    SECRET_KEY = os.environ.get('SECRET_KEY') or 'hard to guess' 
     MAIL_SERVER = os.environ.get('MAIL_SERVER','smtp.googlemail.com')
     MAIL_PORT = int(os.environ.get('MAIL_PORT','587'))
     MAIL_USE_TLS = os.environ.get('MAIL_USE_TLS','true').lower() in ['true','on','1']
@@ -31,7 +31,7 @@ class ProductionConfig(Config):
 
 config = {
     'development' : DevelopmentConfig,
-    'testing' : Testing,
+    'testing' : TestingConfig,
     'production' : ProductionConfig,
     'default' : DevelopmentConfig
 }
