+++ requirements.txt
@@ -1,0 +1,23 @@
+alembic==1.5.2
+blinker==1.4
+click==7.1.2
+dominate==2.6.0
+Flask==1.1.2
+Flask-Bootstrap==3.3.7.1
+Flask-Mail==0.9.1
+Flask-Migrate==2.6.0
+Flask-Moment==0.11.0
+Flask-SQLAlchemy==2.4.4
+Flask-WTF==0.14.3
+itsdangerous==1.1.0
+Jinja2==2.11.2
+Mako==1.1.4
+MarkupSafe==1.1.1
+pkg-resources==0.0.0
+python-dateutil==2.8.1
+python-editor==1.0.4
+six==1.15.0
+SQLAlchemy==1.3.22
+visitor==0.1.3
+Werkzeug==1.0.1
+WTForms==2.3.3
